update browserSync, split code, working on preview
