+++ demo/package.json
@@ -3,7 +3,7 @@
   "type": "module",
   "version": "0.0.0",
   "private": true,
-  "packageManager": "pnpm@8.10.2",
+  "packageManager": "pnpm@8.14.1",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
@@ -11,7 +11,7 @@
   },
   "devDependencies": {
     "typescript": "^5.3.3",
-    "vite": "^5.0.6",
+    "vite": "^5.0.11",
     "vite-plugin-browser-sync": "workspace:*"
   }
 }
