+++ demo/vite.config.ts
@@ -1,5 +1,5 @@
 import VitePluginBrowserSync from 'vite-plugin-browser-sync'
 
 export default {
-  plugins: [VitePluginBrowserSync()],
+  plugins: [VitePluginBrowserSync({ preview: true })],
 }
