+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "vite-plugin-browser-sync",
   "type": "module",
-  "version": "2.0.1",
-  "packageManager": "pnpm@8.11.0",
+  "version": "3.0.0",
+  "packageManager": "pnpm@8.14.1",
   "description": "Add BrowserSync in your Vite project",
   "author": "Applelo",
-  "license": "GPL-3.0",
+  "license": "MIT",
   "homepage": "https://github.com/Applelo/vite-plugin-browser-sync",
   "repository": {
     "type": "git",
@@ -50,17 +50,17 @@
   },
   "dependencies": {
     "@types/browser-sync": "^2.29.0",
-    "browser-sync": "^2.29.3",
+    "browser-sync": "^3.0.2",
     "kolorist": "^1.8.0"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "^2.4.2",
-    "@vitest/coverage-v8": "^1.0.2",
-    "eslint": "^8.55.0",
+    "@antfu/eslint-config": "^2.6.2",
+    "@vitest/coverage-v8": "^1.2.0",
+    "eslint": "^8.56.0",
     "playwright": "^1.40.1",
     "tsup": "^8.0.1",
     "typescript": "^5.3.3",
-    "vite": "^5.0.6",
-    "vitest": "^1.0.2"
+    "vite": "^5.0.11",
+    "vitest": "^1.2.0"
   }
 }
